build changes

+++ build.rake
@@ -12,7 +12,6 @@ task :load_config do
 	@conf = YAML.load_file("config.yml")
 	
 	@deploy_dir = @conf["manager"]["deploydir"]
-	@config_path = @conf["manager"]["configpath"]
 	@libraryPaths = @conf["manager"]["customLibraries"]
 end
 
@@ -37,14 +36,27 @@ namespace :manager do
 	
 	desc "Build the whole shazam"
 	task :build => [:msbuild] do
+		temp_path = Paths.join(@base_dir, "tmp_deploy").to_s;
+		FileUtils.rm_rf temp_path
+		FileUtils.mkdir temp_path
+		
+		unless File.exists? Paths.join(@deploy_dir, 'Web.config').to_s
+			raise "Web.config missing: copy BLocal.Web.Manager.MVC3/web.config.template to the deployment directory, rename to web.config and customize."
+		end
+		
+		FileUtils.cp_r Paths.join(@deploy_dir, 'Web.config').to_s, Paths.join(temp_path, 'Web.config').to_s
 		FileUtils.rm_rf @deploy_dir
 		FileUtils.mkdir @deploy_dir
+		FileUtils.cp_r Paths.join(temp_path, 'Web.config').to_s, Paths.join(@deploy_dir, 'Web.config').to_s
+		
+		FileUtils.rm_rf temp_path
+		
 		FileUtils.cp_r Paths.join(@manager_dir, 'bin').to_s, Paths.join(@deploy_dir, 'bin').to_s
 		FileUtils.cp_r Paths.join(@manager_dir, 'Content').to_s, Paths.join(@deploy_dir, 'Content').to_s
 		FileUtils.cp_r Paths.join(@manager_dir, 'Scripts').to_s, Paths.join(@deploy_dir, 'Scripts').to_s
 		FileUtils.cp_r Paths.join(@manager_dir, 'Views').to_s, Paths.join(@deploy_dir, 'Views').to_s
 		FileUtils.cp_r Paths.join(@manager_dir, 'Global.asax').to_s, Paths.join(@deploy_dir, 'Global.asax').to_s
-		FileUtils.cp_r @config_path, Paths.join(@deploy_dir, 'Web.config').to_s
+		
 		@libraryPaths.each do |libraryPath|
 			libraryPathName = Pathname.new libraryPath
 			FileUtils.cp_r libraryPath, Paths.join(@deploy_dir, 'bin', libraryPathName.basename).to_s
